Move integration-cli/request to internal/test/request…
… and change a bit the method signature
