+++ integration-cli/docker_api_exec_resize_test.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 
 	"github.com/docker/docker/integration-cli/checker"
-	"github.com/docker/docker/integration-cli/request"
+	"github.com/docker/docker/internal/test/request"
 	"github.com/go-check/check"
 )
