+++ integration-cli/docker_deprecated_api_v124_unix_test.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/docker/docker/integration-cli/checker"
-	"github.com/docker/docker/integration-cli/request"
+	"github.com/docker/docker/internal/test/request"
 	"github.com/go-check/check"
 )
