+++ integration/system/info_linux_test.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"testing"
 
-	req "github.com/docker/docker/integration-cli/request"
-	"github.com/docker/docker/integration/internal/request"
+	"github.com/docker/docker/internal/test/request"
+	req "github.com/docker/docker/internal/test/request"
 	"github.com/gotestyourself/gotestyourself/assert"
 	is "github.com/gotestyourself/gotestyourself/assert/cmp"
 	"golang.org/x/net/context"
