+++ integration/system/version_test.go
@@ -3,7 +3,7 @@ package system // import "github.com/docker/docker/integration/system"
 import (
 	"testing"
 
-	"github.com/docker/docker/integration/internal/request"
+	"github.com/docker/docker/internal/test/request"
 	"github.com/gotestyourself/gotestyourself/assert"
 	is "github.com/gotestyourself/gotestyourself/assert/cmp"
 	"golang.org/x/net/context"
