+++ internal/test/request/npipe_windows.go
@@ -1,4 +1,4 @@
-package request // import "github.com/docker/docker/integration-cli/request"
+package request
 
 import (
 	"net"
